fix: Fix Piper detection in dependency checker
The dependency checker was incorrectly reporting Piper as not installed
because it was trying to use --version flag which Piper doesn't support.
When the command failed, it assumed Piper wasn't installed even though
it was found in PATH.

Fixed by using --help flag instead, which Piper does support, to verify
the binary is working properly.